requirements1.txt
streamlit
google-generativeai
langgraph
duckduckgo-search